copy only when user file does not exist

+++ src/configuration.py
@@ -14,7 +14,9 @@ fallback_path = Path(__file__).parent.resolve() / "config.ini"
 if not os.path.exists(datadir):
     os.makedirs(datadir)
 
-shutil.copy(fallback_path, user_path)
+if not os.path.exists(user_path):
+    shutil.copy(fallback_path, user_path)
+
 config_path = user_path
 
 if not config_path.is_file():
